Warn about missing packages, rather than abort
Even with some missing dependencies, an app may work correctly. For
example, the `curl` package could be a dependency of a used Wasm
package, but if network functionality is avoided never actually used at
runtime.

So for missing packages installed locally from a CRAN-like repo, we
issue a warning and only include the package without any Wasm assets.

For GitHub packages we are still strict and require that the package
exists.